requirements.txt
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
colorzero==1.1
Flask==1.0.3
Flask-Cors==3.0.8
google-api-core==1.14.2
google-auth==1.6.3
google-cloud-core==1.0.3
google-cloud-firestore==1.4.0
googleapis-common-protos==1.6.0
gpiozero==1.5.0
grpcio==1.23.0
h11==0.8.1
httptools==0.0.13
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
protobuf==3.9.1
pyasn1==0.4.7
pyasn1-modules==0.2.6
pygame==1.9.6
pytz==2019.2
redis==3.3.8
requests==2.22.0
rsa==4.0
six==1.12.0
urllib3==1.25.3
uvicorn==0.8.6
uvloop==0.12.2
websockets==7.0
Werkzeug==0.15.4
